SASB-99: integrate ECR push into existent workflows
The workflows previously only supported pushing into a `public` docker
registry. With the addition of the docker-ecr-push action, the existing
workflows can support the ECR push.

This is turned off by default, to preserve backwards compatibility, and
requires the access key and secret key to be provided within the
repository secrets.