Make submodule checkout optional
This change adds a boolean flag to enable the checking out of submodules
if applicable. To maintain backwards compatibility, this is set to
default `true`. Future work should endeavour to remove all the checkout
steps and pass that responsibility onto the callee due to the number of
repo specific checkout options are possible.